% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsoilprofile-package.R
\docType{package}
\name{ggsoilprofile-package}
\alias{ggsoilprofile}
\alias{ggsoilprofile-package}
\title{ggsoilprofile: R Package for Stratigraphy of Soil Profile}
\description{
This package is used for stratigraphy plotting of soil profile using ggplot2.
}
\author{
\strong{Maintainer}: Kiwamu Ishikura \email{ishikura.kiwamu@gmail.com}

}
\keyword{internal}
